refactor: extract useElectronIpc hook from App.tsx
Fixes #231 — narrow onWorkspaceReceived catch to JSON.parse only
